feat: Resolver tests (#5308)
