+++ packages/experimental/labs-functions/src/chain/store.ts
@@ -65,7 +65,7 @@ export class ChainStore {
   }
 
   async initialize() {
-    log.info('initializing...', this.info);
+    log('initializing...', this.info);
     try {
       if (this.baseDir && fs.existsSync(this.baseDir)) {
         this._vectorStore = await FaissStore.load(this.baseDir, this._embeddings);
@@ -96,14 +96,14 @@ export class ChainStore {
       );
     }
 
-    log.info('initialized', this.info);
+    log('initialized', this.info);
     return this;
   }
 
   async save() {
     invariant(this.baseDir);
     invariant(this._vectorStore);
-    log.info('saving...', this.info);
+    log('saving...', this.info);
     fs.mkdirSync(this.baseDir, { recursive: true });
     await this._vectorStore.save(this.baseDir);
 
@@ -121,7 +121,7 @@ export class ChainStore {
 
   async delete() {
     invariant(this.baseDir);
-    log.info('deleting...', this.info);
+    log('deleting...', this.info);
     fs.rmSync(this.baseDir, { recursive: true, force: true });
     return this;
   }
