+++ packages/experimental/labs-functions/src/functions/chat/handler.ts
@@ -6,14 +6,11 @@ import { join } from 'node:path';
 
 import { Thread as ThreadType, Message as MessageType } from '@braneframe/types';
 import { sleep } from '@dxos/async';
-import { type Space } from '@dxos/client/echo';
 import { subscriptionHandler } from '@dxos/functions';
-import { log } from '@dxos/log';
 
-import { createContext, createSequence } from './request';
-import { createResolvers, type Resolvers } from './resolvers';
-import { createResponse } from './response';
-import { type ChainResources, type ChainVariant, createChainResources } from '../../chain';
+import { RequestProcessor } from './processor';
+import { createResolvers } from './resolvers';
+import { type ChainVariant, createChainResources } from '../../chain';
 import { getKey } from '../../util';
 
 export const handler = subscriptionHandler(async ({ event, context, response }) => {
@@ -22,15 +19,6 @@ export const handler = subscriptionHandler(async ({ event, context, response }) 
   if (!space || !objects?.length) {
     return response.status(400);
   }
-
-  const config = client.config;
-  const resources = createChainResources((process.env.DX_AI_MODEL as ChainVariant) ?? 'openai', {
-    baseDir: dataDir ? join(dataDir, 'agent/functions/embedding') : undefined,
-    apiKey: getKey(config, 'openai.com/api_key'),
-  });
-  await resources.store.initialize();
-
-  const resolvers = await createResolvers(client.config);
 
   // TODO(burdon): The handler is called before the mutation is processed!
   await sleep(500);
@@ -49,12 +37,22 @@ export const handler = subscriptionHandler(async ({ event, context, response }) 
 
   // Process threads.
   if (activeThreads) {
+    const resources = createChainResources((process.env.DX_AI_MODEL as ChainVariant) ?? 'openai', {
+      baseDir: dataDir ? join(dataDir, 'agent/functions/embedding') : undefined,
+      apiKey: getKey(client.config, 'openai.com/api_key'),
+    });
+
+    await resources.store.initialize();
+    const resolvers = await createResolvers(client.config);
+
+    const processor = new RequestProcessor(resources, resolvers);
+
     await Promise.all(
       Array.from(activeThreads).map(async (thread) => {
         const message = thread.messages[thread.messages.length - 1];
         if (message.__meta.keys.length === 0) {
-          const blocks = await processMessage(resources, resolvers, space, thread, message);
-          if (blocks) {
+          const blocks = await processor.processThread(space, thread, message);
+          if (blocks?.length) {
             thread.messages.push(
               new MessageType(
                 {
@@ -65,7 +63,7 @@ export const handler = subscriptionHandler(async ({ event, context, response }) 
                 },
                 {
                   meta: {
-                    keys: [{ source: 'openai.com' }],
+                    keys: [{ source: 'openai.com' }], // TODO(burdon): Get from chain resources.
                   },
                 },
               ),
@@ -76,70 +74,3 @@ export const handler = subscriptionHandler(async ({ event, context, response }) 
     );
   }
 });
-
-// TODO(burdon): Create class.
-const processMessage = async (
-  resources: ChainResources,
-  resolvers: Resolvers,
-  space: Space,
-  thread: ThreadType,
-  message: MessageType,
-): Promise<MessageType.Block[] | undefined> => {
-  let blocks: MessageType.Block[] | undefined;
-  const { start, stop } = createStatusNotifier(space, thread.id);
-  try {
-    const text = message.blocks
-      .map((message) => message.text)
-      .filter(Boolean)
-      .join('\n');
-
-    const match = text.match(/\/(\w+)\s*(.+)?/);
-    if (match) {
-      const prompt = match[1];
-      const content = match[2];
-
-      start();
-      log.info('processing', { prompt, content });
-      const context = createContext(space, message, thread);
-      const sequence = await createSequence(space, resources, context, resolvers, { prompt });
-      const response = await sequence.invoke(content);
-
-      blocks = createResponse(space, context, response);
-      log.info('response', { blocks });
-    }
-  } catch (err) {
-    log.error('processing message', err);
-    blocks = [{ text: 'Error generating response.' }];
-  } finally {
-    stop();
-  }
-
-  return blocks;
-};
-
-// TODO(burdon): Factor out.
-const createStatusNotifier = (space: Space, id: string) => {
-  let start: number | undefined;
-  return {
-    start: () => {
-      if (!start) {
-        start = Date.now();
-        void space.postMessage(`status/${id}`, {
-          event: 'processing',
-          ts: start,
-        });
-      }
-    },
-    stop: () => {
-      if (start) {
-        const now = Date.now();
-        void space.postMessage(`status/${id}`, {
-          event: 'done',
-          ts: now,
-          duration: now - start,
-        });
-        start = undefined;
-      }
-    },
-  };
-};
