+++ packages/experimental/labs-functions/src/functions/chat/parser.ts
@@ -3,7 +3,8 @@
 //
 
 import { invariant } from '@dxos/invariant';
-import { log } from '@dxos/log';
+
+import { safeParseJson } from '../../util';
 
 export type ParseResult = {
   timestamp: string;
@@ -21,21 +22,20 @@ export const parseMessage = (content: string, type?: string): ParseResult => {
 
   // Check if raw JSON.
   if (!type || type === 'json') {
-    const value = parseJson(content);
+    const value = safeParseJson(content);
     if (value) {
       return {
         timestamp,
         type: 'json',
-        content: value,
+        content,
         data: value,
       };
     }
   }
 
-  // Check for embedded block content.
+  // Check for fenced content.
   const regexp = new RegExp('(.+)?```\\s*(' + (type ?? '\\w+') + ')?\\s+(.+)```', 's');
   const match = regexp.exec(content);
-  log.info('match', { match });
   if (match) {
     const [_, pre, type, content, post] = match;
     return {
@@ -44,7 +44,7 @@ export const parseMessage = (content: string, type?: string): ParseResult => {
       pre,
       post,
       content,
-      data: type === 'json' ? parseJson(content) : undefined,
+      data: type === 'json' ? safeParseJson(content) : undefined,
       kind: 'fenced',
     };
   }
@@ -54,12 +54,4 @@ export const parseMessage = (content: string, type?: string): ParseResult => {
     type: 'text',
     content,
   };
-};
-
-export const parseJson = (content: string) => {
-  try {
-    return JSON.parse(content);
-  } catch (err) {
-    return null;
-  }
 };
