+++ packages/experimental/labs-functions/src/functions/chat/resolvers/type.ts
@@ -1,0 +1,9 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export type Resolver = () => Promise<any>;
+
+export type ResolverMap = {
+  [key: string]: Resolver | ResolverMap;
+};
